backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
requests==2.32.5
cryptography==45.0.7
python-dotenv==1.1.1
langgraph==0.6.6
langchain-openai==0.3.33
langchain-core==0.3.76
pydantic==2.11.9
